+++ CHANGELOG.md
@@ -1,3 +1,4 @@
+# v0.3.0
 - add: creation time display
 - add: sort by creation time
 - add: randomly select